+++ CHANGELOG.md
@@ -13,6 +13,7 @@ develop
 
 ### Fixes
 
+- Fix an issue where the regular expression didn't handle `\v` metacharacter correctly.
 - Fix an issue where the selection of encoding menu in toolbar didn't restore to the previous one when encoding reinterpretation was failed.
 - [AppStore ver.] Fix an issue where acknowledgement window was empty.